// src/types.ts
export type ResourceID = string;
export type CategoryID = string;
export type EncounterID = string;
export type EnemyID = string;
export type AllyID = string;
export type ActionID = string;
export type ItemID = string;
export type SkillID = string;
export type PartyUnitID = string;
export type EnemyUnitID = number;
export type ProgressionID = string;

export const MainCharacterPartyUnitID = "player";
export type ThemeID = 'light' | 'dark' | 'system';
export type Tab = 'actions' | 'encounter' | 'inventory' | 'config';
export const DefaultTab = 'actions';

export interface CategoryConfig {
    id: CategoryID;
    name: string;
    description: string;
}

export interface Prerequisite {
    // Resource prerequisites
    resourceID?: ResourceID;
    minValue?: number;
    maxValue?: number;
    minMax?: number;
    maxMax?: number;
    valueGreaterThanZero?: boolean;
    valueLessThanMax?: boolean;
    valueEqualToMax?: boolean;
    // Action prerequisites
    actionID?: ActionID;
    actionMinExecutions?: number;
    actionMaxExecutions?: number;
    // Unit prerequisites
    unit?: 'mainCharacter' | 'enemy' | 'enemies' | 'ally' | 'allies' | 'all';
    unitStat?: 'health' | 'mana';
    // Progression prerequisites
    progressionID?: ProgressionID;
    metaProgressionID?: ProgressionID;
}
// ...
export interface ActionConfig {
    id: ActionID;
    name: string;
    category: CategoryID;
    description: string;
    flavorText: string;
    globalCosts?: GlobalCost[];
    unitCosts?: UnitCost[];
    globalEffects: GlobalEffectConfig[];
    unitEffects: UnitEffectConfig[];
    prerequisites: Prerequisite[];
    maxExecutions?: number;
}

export interface ActionState {
    visible: boolean;
    executions: number;
}

export interface GlobalEffectConfig {
    type: 'increase_value' | 'increase_max' | 'start_encounter' | 'end_encounter' | 'enable_progression' | 'enable_meta_progression' | 'loop';
    amount?: number;
    actionID?: ActionID;
    resourceID?: ResourceID;
    encounterID?: EncounterID;
    itemID?: ItemID;
    progressionID?: ProgressionID;
    metaProgressionID?: ProgressionID;

    chance?: number;
    hidden?: boolean;
}

export interface UnitEffectConfig {
    type: 'increase_value' | 'increase_max' | 'damage';
    target: 'mainCharacter' | 'ally' | 'allies' | 'enemy' | 'enemies' | 'all';
    stat: 'health' | 'mana';
    amount: number;
}
// ...
export interface GlobalCost {
    resourceID: ResourceID;
    amount: number;
}

export interface UnitCost {
    target: 'mainCharacter' | 'self' | 'ally' | 'allies' | 'enemy' | 'enemies' | 'all';
    stat: 'health' | 'mana';
    amount: number;
}